trace: regenerate trace from go@d3d4e7b
This includes CL 601455, which reduces the size of events, and
CL 608855, which refactors Stack.Frames to return an iter.Seq.
This also includes CL 608856, which refactors some of the trace parser
to use the range-over-func version of Stack.Frames, and as such requires
bumping the go:build directive to go1.23.

This also includes CL 593975, which merged some debugging tools in to
the main cmd/trace. So, those tools are removed from this repo.

[git-generate]
cd trace
./gen.bash d3d4e7b0d10d5587093b769ed600bd3b0a3fd9c1

Change-Id: I4987c031461a4a6458cc5e3e0db7f05e993120fb